+++ examples/spring-boot-2.4/test.http
@@ -1,0 +1,39 @@
+### Create Book: Hamlet
+POST http://localhost:8080/api/v1/books
+Content-Type: application/json
+
+{
+  "name": "Hamlet"
+}
+
+### Create Book: King Lear
+POST http://localhost:8080/api/v1/books
+Content-Type: application/json
+
+{
+"name": "King Lear"
+}
+
+### Create Book: Romeo and Juliet
+POST http://localhost:8080/api/v1/books
+Content-Type: application/json
+
+{
+  "name": "Romeo and Juliet"
+}
+
+### Create Book: Macbeth
+POST http://localhost:8080/api/v1/books
+Content-Type: application/json
+
+{
+  "name": "Macbeth"
+}
+
+### Find Book by id
+GET localhost:8080/api/v1/books/1
+Content-Type: application/json
+
+### Find Book by name
+GET localhost:8080/api/v1/books?name=et
+Content-Type: application/json
