+++ examples/spring-boot-2.5/src/main/kotlin/com/linecorp/kotlinjdsl/spring/data/example/Main.kt
@@ -1,0 +1,12 @@
+package com.linecorp.kotlinjdsl.spring.data.example
+
+import org.springframework.boot.autoconfigure.SpringBootApplication
+import org.springframework.boot.runApplication
+import org.springframework.web.bind.annotation.*
+
+@SpringBootApplication
+class ExampleApplication
+
+fun main(args: Array<String>) {
+    runApplication<ExampleApplication>(*args)
+}
