+++ examples/spring-boot-2.5/src/main/kotlin/com/linecorp/kotlinjdsl/spring/data/example/entity/Book.kt
@@ -1,0 +1,17 @@
+package com.linecorp.kotlinjdsl.spring.data.example.entity
+
+import javax.persistence.*
+
+@Entity
+@Table(name = "book")
+data class Book(
+    @Id
+    @Column(name = "id")
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    val id: Long = 0,
+
+    @Column
+    val name: String,
+) {
+    fun toJson() = """{"id":$id,"name":"$name"} """
+}
