+++ examples/spring-boot-2.6/build.gradle.kts
@@ -1,0 +1,22 @@
+plugins {
+    id("org.springframework.boot") version Dependencies.springBootVersion
+    kotlin("plugin.spring")
+    kotlin("plugin.jpa")
+}
+
+apply(plugin = "org.springframework.boot")
+apply(plugin = "kotlin-spring")
+apply(plugin = "kotlin-jpa")
+
+dependencies {
+    // implementation("com.linecorp.kotlin-jdsl:spring-data-kotlin-jdsl-starter:x.y.z")
+    implementation(Modules.springDataStarter)
+
+    implementation(Dependencies.springBootWeb)
+    implementation(Dependencies.springBootJpa)
+    implementation(Dependencies.jacksonKotlinModule)
+    implementation(Dependencies.h2)
+    implementation(platform(Dependencies.springBootBom))
+
+    testImplementation(Dependencies.springBootTest)
+}
