--- examples/spring-boot/src/main/kotlin/com/linecorp/kotlinjdsl/spring/data/example/entity/Book.kt
@@ -1,15 +1,0 @@
-package com.linecorp.kotlinjdsl.spring.data.example.entity
-
-import javax.persistence.*
-
-@Entity
-@Table(name = "book")
-data class Book(
-    @Id
-    @Column(name = "id")
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    val id: Long = 0,
-
-    @Column
-    val name: String,
-)
